Added google sign in, public folder, improved update user settings, improved isAdmin middleware, refactored some code

+++ src/controllers/users/delete.controller.js
@@ -4,6 +4,10 @@ module.exports.deleteOneById = (req, res) => {
   try {
     const { id } = req.params;
 
+    if (!req.isAdmin) {
+      return res.status(401).send("Must be admin");
+    }
+
     return deleteOneById(id)
       .then((usr) => {
         return res.status(200).send(usr);
